Adds accessibility for Pie Charts. (#1060)
This commit makes PieChartView adhere to the UIAccessibility container
protocol. See the Accessibility section marked for the methods. To allow
conformance, 3 properties are added to IPieChartDataSet and
PieChartDataSet that enable a more user friendly audio descriptionfor
data elements. PieChartRenderer has a new private method
createAccessibleElement() that populates the accessiblePieChartElements
property, which in turn is used in PieChartView. Note that to prevent
contextless audio descriptions, the default Chart Description header
text was deleted in Description.swift since it is already an optional.